backend/requirements.txt
fastapi
uvicorn
pydantic
websockets==12.0
python-dotenv==1.0.1
pymongo==4.7.2
mongomock==4.1.2
redis==5.0.4
httpx==0.27.0
pytest==8.2.0
groq==0.8.0
deepgram-sdk==3.4.0
cartesia
